Rename Feature module to Eon
