+++ cardano-api/internal/Cardano/Api/Eon/ConwayEraOnwards.hs
@@ -8,15 +8,15 @@
 {-# LANGUAGE TypeFamilies #-}
 {-# LANGUAGE TypeOperators #-}
 
-module Cardano.Api.Feature.BabbageEraOnwards
-  ( BabbageEraOnwards(..)
-  , IsBabbageEraOnwards(..)
-  , AnyBabbageEraOnwards(..)
-  , babbageEraOnwardsConstraints
-  , babbageEraOnwardsToCardanoEra
-  , babbageEraOnwardsToShelleyBasedEra
+module Cardano.Api.Eon.ConwayEraOnwards
+  ( ConwayEraOnwards(..)
+  , IsConwayEraOnwards(..)
+  , AnyConwayEraOnwards(..)
+  , conwayEraOnwardsConstraints
+  , conwayEraOnwardsToCardanoEra
+  , conwayEraOnwardsToShelleyBasedEra
 
-  , BabbageEraOnwardsConstraints
+  , ConwayEraOnwardsConstraints
   ) where
 
 import           Cardano.Api.Eras.Core
@@ -29,7 +29,9 @@ import qualified Cardano.Crypto.Hash.Class as C
 import qualified Cardano.Crypto.VRF as C
 import qualified Cardano.Ledger.Api as L
 import qualified Cardano.Ledger.BaseTypes as L
-import qualified Cardano.Ledger.Core as L
+import qualified Cardano.Ledger.Conway.Core as L
+import qualified Cardano.Ledger.Conway.Governance as L
+import qualified Cardano.Ledger.Conway.TxCert as L
 import qualified Cardano.Ledger.SafeHash as L
 import qualified Ouroboros.Consensus.Protocol.Abstract as Consensus
 import qualified Ouroboros.Consensus.Protocol.Praos.Common as Consensus
@@ -38,82 +40,78 @@ import qualified Ouroboros.Consensus.Shelley.Ledger as Consensus
 import           Data.Aeson
 import           Data.Typeable (Typeable)
 
-class IsShelleyBasedEra era => IsBabbageEraOnwards era where
-  babbageEraOnwards :: BabbageEraOnwards era
+class IsShelleyBasedEra era => IsConwayEraOnwards era where
+  conwayEraOnwards :: ConwayEraOnwards era
 
-data BabbageEraOnwards era where
-  BabbageEraOnwardsBabbage :: BabbageEraOnwards BabbageEra
-  BabbageEraOnwardsConway  :: BabbageEraOnwards ConwayEra
+data ConwayEraOnwards era where
+  ConwayEraOnwardsConway :: ConwayEraOnwards ConwayEra
 
-deriving instance Show (BabbageEraOnwards era)
-deriving instance Eq (BabbageEraOnwards era)
+deriving instance Show (ConwayEraOnwards era)
+deriving instance Eq (ConwayEraOnwards era)
 
-instance IsBabbageEraOnwards BabbageEra where
-  babbageEraOnwards = BabbageEraOnwardsBabbage
+instance IsConwayEraOnwards ConwayEra where
+  conwayEraOnwards = ConwayEraOnwardsConway
 
-instance IsBabbageEraOnwards ConwayEra where
-  babbageEraOnwards = BabbageEraOnwardsConway
-
-instance FeatureInEra BabbageEraOnwards where
+instance FeatureInEra ConwayEraOnwards where
   featureInEra no yes = \case
     ByronEra    -> no
     ShelleyEra  -> no
     AllegraEra  -> no
     MaryEra     -> no
     AlonzoEra   -> no
-    BabbageEra  -> yes BabbageEraOnwardsBabbage
-    ConwayEra   -> yes BabbageEraOnwardsConway
+    BabbageEra  -> no
+    ConwayEra   -> yes ConwayEraOnwardsConway
 
-instance ToCardanoEra BabbageEraOnwards where
+instance ToCardanoEra ConwayEraOnwards where
   toCardanoEra = \case
-    BabbageEraOnwardsBabbage -> BabbageEra
-    BabbageEraOnwardsConway  -> ConwayEra
+    ConwayEraOnwardsConway -> ConwayEra
 
-data AnyBabbageEraOnwards where
-  AnyBabbageEraOnwards :: BabbageEraOnwards era -> AnyBabbageEraOnwards
+data AnyConwayEraOnwards where
+  AnyConwayEraOnwards :: ConwayEraOnwards era -> AnyConwayEraOnwards
 
-deriving instance Show AnyBabbageEraOnwards
+deriving instance Show AnyConwayEraOnwards
 
-type BabbageEraOnwardsConstraints era =
+type ConwayEraOnwardsConstraints era =
   ( C.HashAlgorithm (L.HASH (L.EraCrypto (ShelleyLedgerEra era)))
   , C.Signable (L.VRF (L.EraCrypto (ShelleyLedgerEra era))) L.Seed
   , Consensus.PraosProtocolSupportsNode (ConsensusProtocol era)
   , Consensus.ShelleyCompatible (ConsensusProtocol era) (ShelleyLedgerEra era)
   , L.ADDRHASH (Consensus.PraosProtocolSupportsNodeCrypto (ConsensusProtocol era)) ~ Blake2b.Blake2b_224
-  , L.BabbageEraPParams (ShelleyLedgerEra era)
-  , L.BabbageEraTxBody (ShelleyLedgerEra era)
-  , L.BabbageEraTxOut (ShelleyLedgerEra era)
+  , L.ConwayEraGov (ShelleyLedgerEra era)
+  , L.ConwayEraPParams (ShelleyLedgerEra era)
+  , L.ConwayEraTxBody (ShelleyLedgerEra era)
+  , L.ConwayEraTxCert (ShelleyLedgerEra era)
   , L.Crypto (L.EraCrypto (ShelleyLedgerEra era))
   , L.Era (ShelleyLedgerEra era)
   , L.EraCrypto (ShelleyLedgerEra era) ~ L.StandardCrypto
+  , L.EraGov (ShelleyLedgerEra era)
   , L.EraPParams (ShelleyLedgerEra era)
   , L.EraTx (ShelleyLedgerEra era)
   , L.EraTxBody (ShelleyLedgerEra era)
   , L.HashAnnotated (L.TxBody (ShelleyLedgerEra era)) L.EraIndependentTxBody L.StandardCrypto
   , L.ShelleyEraTxBody (ShelleyLedgerEra era)
   , L.ShelleyEraTxCert (ShelleyLedgerEra era)
+  , L.TxCert (ShelleyLedgerEra era) ~ L.ConwayTxCert (ShelleyLedgerEra era)
 
   , FromCBOR (Consensus.ChainDepState (ConsensusProtocol era))
   , FromCBOR (DebugLedgerState era)
-  , IsBabbageEraOnwards era
   , IsCardanoEra era
   , IsShelleyBasedEra era
+  , IsConwayEraOnwards era
   , ToJSON (DebugLedgerState era)
   , Typeable era
   )
 
-babbageEraOnwardsConstraints :: ()
-  => BabbageEraOnwards era
-  -> (BabbageEraOnwardsConstraints era => a)
+conwayEraOnwardsConstraints :: ()
+  => ConwayEraOnwards era
+  -> (ConwayEraOnwardsConstraints era => a)
   -> a
-babbageEraOnwardsConstraints = \case
-  BabbageEraOnwardsBabbage -> id
-  BabbageEraOnwardsConway  -> id
+conwayEraOnwardsConstraints = \case
+  ConwayEraOnwardsConway -> id
 
-babbageEraOnwardsToCardanoEra :: BabbageEraOnwards era -> CardanoEra era
-babbageEraOnwardsToCardanoEra = shelleyBasedToCardanoEra . babbageEraOnwardsToShelleyBasedEra
+conwayEraOnwardsToCardanoEra :: ConwayEraOnwards era -> CardanoEra era
+conwayEraOnwardsToCardanoEra = shelleyBasedToCardanoEra . conwayEraOnwardsToShelleyBasedEra
 
-babbageEraOnwardsToShelleyBasedEra :: BabbageEraOnwards era -> ShelleyBasedEra era
-babbageEraOnwardsToShelleyBasedEra = \case
-  BabbageEraOnwardsBabbage -> ShelleyBasedEraBabbage
-  BabbageEraOnwardsConway  -> ShelleyBasedEraConway
+conwayEraOnwardsToShelleyBasedEra :: ConwayEraOnwards era -> ShelleyBasedEra era
+conwayEraOnwardsToShelleyBasedEra = \case
+  ConwayEraOnwardsConway -> ShelleyBasedEraConway
