+++ cardano-api/internal/Cardano/Api/Eon/ShelleyToMaryEra.hs
@@ -8,7 +8,7 @@
 {-# LANGUAGE TypeFamilies #-}
 {-# LANGUAGE TypeOperators #-}
 
-module Cardano.Api.Feature.ShelleyToMaryEra
+module Cardano.Api.Eon.ShelleyToMaryEra
   ( ShelleyToMaryEra(..)
   , IsShelleyToMaryEra(..)
   , AnyShelleyToMaryEra(..)
